+++ test/integration/run.js
@@ -1,0 +1,18 @@
+const path = require('path');
+const { runTests } = require('@vscode/test-electron');
+
+async function main() {
+  const extensionDevelopmentPath = path.resolve(__dirname, '../../');
+  const extensionTestsPath = path.resolve(__dirname, './index');
+
+  await runTests({
+    extensionDevelopmentPath,
+    extensionTestsPath,
+    launchArgs: ['--disable-extensions'],
+  });
+}
+
+main().catch((err) => {
+  console.error('Failed to run tests:', err);
+  process.exit(1);
+});